Rewrite feature test using cheerio
Cheerio is an implementation of core jQuery functionality, which can be
used to parse an html page. So I can rewrite my expectation that the
page will include "<h1>Dishwashers</h1>" to expect instead that the page
will have an <h1> element containing the text "Dishwashers". This makes
the test more robust, and less tightly coupled to the view.