Create a ProductList class
The ProductList constructor takes the API response as an input, and
converts it into an array of ProductSummary objects which can be passed
into the view.

At this point the API started giving me a "Sorry we can't find the page
you've requested" error" (see `api-bad-response.html` in the project
root). I don't think I changed anything in my code to cause this, so it
could be something on the API's side, but for now I've added a helper
method which checks to see if a valid response came back and if not
substitute the saved data I've been using for testing. I'll have to
investigate this further later.